refactor: remove zod and move files to anothe folder
